refactor(engine): move BlockOrPayload to tree types
